+++ src/ipfs/config.ts
@@ -24,6 +24,7 @@ export function iframe(): Promise<MessagePort> {
     iframe.style.width = "0"
     iframe.style.height = "0"
     iframe.style.border = "none"
+    iframe.style.display = "none"
     document.body.appendChild(iframe)
 
     iframe.onload = () => {